Driver for three-layer, core-adjusting model.
Run with
  python drive_3l_model.py planet y1 z1 z2
or
  python drive_3l_model.py --help
to list required and optional parameters.
Remember, this model _specifies_ y1, z1, and z2, in addition to
planet's mass and radius (read from observables struct) and mean y
(default 0.275), as well as radius of envelope inner/outer boundary
(default 0.8 but probably worth adjusting per-case). The core radius
will be adjusted by the model, although an initial guess can be
given to help with convergence.